requirements_A100.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==4.0.3
attrs==20.3.0
backcall==0.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==3.2.3
brotlipy==0.7.0
cachetools==5.3.2
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==3.3.1
click==8.1.7
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
datasets==2.14.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.7
dnspython==2.1.0
entrypoints==0.3
evaluate==0.4.1
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
huggingface-hub==0.17.3
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata==6.8.0
ipykernel==5.4.3
ipython @ file:///tmp/build/80754af9/ipython_1610724799192/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
jiwer==3.0.3
joblib==1.3.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
lightning-utilities==0.9.0
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==10.1.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
packaging==23.2
pandas==2.0.3
pandocfilters==1.4.3
parso==0.7.0
path==16.7.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pkginfo==1.7.0
prometheus-client==0.9.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==4.24.4
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyDeprecate==0.3.1
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytorch-lightning==1.4.2
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.3.1
pyzmq==22.0.0
rapidfuzz==3.4.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
ruamel_yaml==0.15.87
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.14.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchelastic==0.2.1
torchmetrics==0.5.0
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.66.1
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
transformers==4.34.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
